Create models.py
This code is a class definition in Python. The code defines two classes: user and student. The user class has two attributes, name and email, and a method called send_email that checks if the email attribute is not null and then prints a "sending email" message. The student class inherits from the user class and adds an additional attribute called id. The __init__ method of the student class calls the __init__ method of the user class using the super() function and assigns a unique value generated by uuid.uuid4() to the id attribute.

If you're interested in learning more about Python programming, I would recommend the "Python Practical Course: Creating a CRUD" course on Platzi. In this course, you'll learn how to create CRUD applications using Python. You can also consider the "Design and Layout Course for Emails" and the "Proper Configuration Course in Email Marketing"